Major update (breaking change)
:white_check_mark: Coding Style.
:white_check_mark: Refactored the Core file into separate classes.
:white_check_mark: Laravel Package Discovery.
:white_check_mark: Alias( Facade).
:white_check_mark: Using Laravel Log for paytabs logs.
:white_check_mark: Removed `routes.php` file and added it manually (there is a line that can be used if `routes.php` is needed).
