--- routes/routes.php
@@ -1,3 +1,0 @@
-<?php
-
-Route::post('/paymentIPN', [\Paytabscom\Laravel_paytabs\Controllers\PaytabsLaravelListenerApi::class, 'paymentIPN'])->name('payment_ipn');
