+++ src/Services/IpnRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Paytabscom\Laravel_paytabs\Services;
+namespace Paytabscom\LaravelPaytabs\Services;
 
 use Illuminate\Http\Request;
 
@@ -43,9 +43,8 @@ class IpnRequest
     private function isValidIPNRequest($httpRequest){
         $signature= $httpRequest->header('signature');
         $content= $httpRequest->getContent(); //get the request raw content
-        new \Paytabscom\Laravel_paytabs\paytabs_core(); //this is a hack just to be able to use `PaytabsApi` class from paytabs_core.php !!!
 
-        $paytabs_api= \Paytabscom\Laravel_paytabs\PaytabsApi::getInstance(config('paytabs.region'), config('paytabs.profile_id'), config('paytabs.server_key'));
+        $paytabs_api= \Paytabscom\LaravelPaytabs\Core\PaytabsApi::getInstance(config('paytabs.region'), config('paytabs.profile_id'), config('paytabs.server_key'));
         if($paytabs_api->is_valid_ipn($content, $signature)){
             return true;
         }else{
