+++ src/core/PaytabsHelper.php
@@ -1,0 +1,161 @@
+<?php
+
+namespace Paytabscom\LaravelPaytabs\Core;
+
+use Illuminate\Support\Facades\Log;
+
+/**
+ * PayTabs v2 PHP SDK
+ * Version: 2.0.8
+ */
+abstract class PaytabsHelper
+{
+    static function paymentType($key)
+    {
+        return PaytabsApi::PAYMENT_TYPES[$key]['name'];
+    }
+
+    static function paymentAllowed($code, $currencyCode)
+    {
+        $row = null;
+        foreach (PaytabsApi::PAYMENT_TYPES as $key => $value) {
+            if ($value['name'] === $code) {
+                $row = $value;
+                break;
+            }
+        }
+        if (!$row) {
+            return false;
+        }
+        $list = $row['currencies'];
+        if ($list == null) {
+            return true;
+        }
+
+        $currencyCode = strtoupper($currencyCode);
+
+        return in_array($currencyCode, $list);
+    }
+
+    static function isPayTabsPayment($code)
+    {
+        foreach (PaytabsApi::PAYMENT_TYPES as $key => $value) {
+            if ($value['name'] === $code) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * @return the first non-empty var from the vars list
+     * @return null if all params are empty
+     */
+    public static function getNonEmpty(...$vars)
+    {
+        foreach ($vars as $var) {
+            if (!empty($var)) return $var;
+        }
+        return null;
+    }
+
+    /**
+     * convert non-english digits to English
+     * used for fileds that accepts only English digits like: "postal_code"
+     */
+    public static function convertAr2En($string)
+    {
+        $nonEnglish = [
+            // Arabic
+            [
+                '٠',
+                '١',
+                '٢',
+                '٣',
+                '٤',
+                '٥',
+                '٦',
+                '٧',
+                '٨',
+                '٩'
+            ],
+            // Persian
+            [
+                '۰',
+                '۱',
+                '۲',
+                '۳',
+                '۴',
+                '۵',
+                '۶',
+                '۷',
+                '۸',
+                '۹'
+            ]
+        ];
+
+        $num = range(0, 9);
+
+        $englishNumbersOnly = $string;
+        foreach ($nonEnglish as $oldNum) {
+            $englishNumbersOnly = str_replace($oldNum, $num, $englishNumbersOnly);
+        }
+
+        return $englishNumbersOnly;
+    }
+
+    /**
+     * check Strings that require to be a valid Word, not [. (dot) or digits ...]
+     * if the parameter is not a valid "Word", convert it to "NA"
+     */
+    public static function pt_fillIfEmpty(&$string)
+    {
+        if (empty(preg_replace('/[\W]/', '', $string))) {
+            $string .= 'NA';
+        }
+    }
+
+    static function pt_fillIP(&$string)
+    {
+        $string = $_SERVER['REMOTE_ADDR'];
+    }
+
+    /**
+     * <b>paytabs_error_log<b> should be defined,
+     * Main functionality: use the platform logger to log the error messages
+     * If not found: create a new log file and log the messages
+     */
+    public static function log($msg)
+    {
+        try {
+            Log::channel('paytabs')->info($msg);
+        } catch (\Throwable $th) {
+            try {
+                $_prefix = date('c') . ' PayTabs: ';
+                $_msg = ($_prefix . $msg . PHP_EOL);
+                file_put_contents('debug_paytabs.log', $_msg, FILE_APPEND);
+            } catch (\Throwable $th) {
+                // var_export($th);
+            }
+        }
+    }
+
+    static function getTokenInfo($return_values)
+    {
+        $fields = [
+            'pt_token',
+            'pt_customer_email',
+            'pt_customer_password'
+        ];
+
+        $tokenInfo = [];
+
+        foreach ($fields as $field) {
+            if (!isset($return_values[$field])) return false;
+            $tokenInfo[$field] = $return_values[$field];
+        }
+
+        return $tokenInfo;
+    }
+
+}
